Addressed pr review changes. Updated email services to send email only. Updated user verification to be part of auth
