+++ src/auth/auth.controller.ts
@@ -1,4 +1,13 @@
-import { Body, Controller, HttpCode, HttpStatus, Post } from '@nestjs/common';
+import {
+  BadRequestException,
+  Body,
+  Controller,
+  Get,
+  HttpCode,
+  HttpStatus,
+  Post,
+  Query,
+} from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { SignupRequestData } from './dto/signup-user-dto';
 
@@ -10,5 +19,18 @@ export class AuthController {
   @Post('sign-up')
   async signup(@Body() signupRequestData: SignupRequestData) {
     return this.authService.signUp(signupRequestData);
+  }
+
+  @Post('resend-verification')
+  async reSendVerification(@Body('email') email: string) {
+    if (!email) throw new BadRequestException('Email is required');
+
+    return await this.authService.sendVerificationLink(email);
+  }
+
+  @Get('verify-email')
+  async verifyEmail(@Query('token') token: string) {
+    if (!token) throw new BadRequestException('Token is required');
+    return await this.authService.verify(token);
   }
 }
