+++ src/auth/auth.module.ts
@@ -2,13 +2,15 @@ import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { JwtModule } from '@nestjs/jwt';
-import { UserModule } from 'src/user/user.module';
 import { HashService } from './hash.service';
 import { EmailModule } from 'src/email/email.module';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { User } from 'src/user/user.entity';
+import { EmailVerification } from 'src/auth/email-verification.entity';
 
 @Module({
   imports: [
-    UserModule,
+    TypeOrmModule.forFeature([User, EmailVerification]),
     EmailModule,
     JwtModule.register({
       global: true,
