+++ src/auth/interface.ts
