+++ src/email/email.module.ts
@@ -1,13 +1,12 @@
 import { Module } from '@nestjs/common';
 import { EmailService } from './email.service';
-import { EmailController } from './email.controller';
-import { UserModule } from 'src/user/user.module';
+
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { EmailVerification } from './email-verification.entity';
+import { EmailVerification } from '../auth/email-verification.entity';
 
 @Module({
-  imports: [UserModule, TypeOrmModule.forFeature([EmailVerification])],
-  controllers: [EmailController],
+  imports: [TypeOrmModule.forFeature([EmailVerification])],
+  controllers: [],
   providers: [EmailService],
   exports: [EmailService],
 })
